[lib] Refactor logout in keyserver reducer
Summary:
issue: [[ https://linear.app/comm/issue/ENG-5826/remove-from-keyserverstore-only-the-keyserver-that-was-logged-out-of | ENG-5826 ]]
We want to only remove the keyservers that the user has disconnected from. The only exception is Ashoat's keyserver, which always needs to be in the store. Additionally to what we were removing from it before I removed `lateResponses`. `lateResponses` include message ids, so I'm not sure if it would be ba security breach to leave them. But it looks like this field is only used for the disconnected bar, which is not shown on the logout screen anyway, and soon will not be even connected to the keyserver connection, but to the internet connection ([[https://linear.app/comm/issue/ENG-6137/modify-disconnectedbar-so-it-pulls-from-connectivityinfo | ENG-6137 ]])

Test Plan: Ran the provided test

Reviewers: michal, kamil, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10658