[lib] Update `useResolvedThreadInfos` to remove `LegacyThreadInfo`
Summary:
As part of https://linear.app/comm/issue/ENG-6366/rename-minimallyencodedrawthreadinfo-to-rawthreadinfo

We want to remove `LegacyResolvedThreadInfo` altogether.

To start we're updating `useResolvedThreadInfos` (and related functions) to only handle `MinimallyEncodedThreadInfo`s.

This diff makes that change and addresses the `flow` issues that arose in `lib`. Will create separate diffs for `web` and `native`, the CI will fail for this diff. Personally would've preferred to include all of them in a single diff that passes CI, but will break it up.

[skip-ci]

---

Depends on D10655

Test Plan:
`flow`

Reviewers:
ashoat, ginsu, tomek