[native] Remove instances of `LegacyThreadInfo` throughout `native`
Summary:
Part of https://linear.app/comm/issue/ENG-6366/rename-minimallyencodedrawthreadinfo-to-rawthreadinfo

This diff addresses `flow` issues that arose in `native` as a result of D10670. There are more instances of this type that will be removed in subsequent diffs. This diff just gets us back to 0 flow issues.

---

Depends on D10671

Test Plan: `flow`, CI will run for this diff

Reviewers: ashoat, ginsu, tomek

Differential Revision: https://phab.comm.dev/D10672