[keyserver] Use verifyUserLoggedIn instead of fetchIdentityInfo in versionResponseValidator
Summary:
`verifyUserLoggedIn` returns the same thing as `fetchIdentityInfo`. The difference is that `verifyUserLoggedIn` will attempt identity login if the identity info is not present in MariaDB.

This makes it a bit more convenient to be able to eg. clear the identity info from MariaDB to force a re-login, which came up in [ENG-6466](https://linear.app/comm/issue/ENG-6466/ashoats-keyserver-not-using-userid-256).

Test Plan:
1. Flow
2. Careful reading of [the code](https://github.com/CommE2E/comm/blob/a53d9f43351c977111039c46d9fa356556069652/keyserver/src/user/login.js#L57C16-L67) for `verifyUserLoggedIn`
3. An observation of what happened when I restarted my keyserver after clearing the identity info from MariaDB (restart triggered running `verifyUserLoggedIn`)

Reviewers: varun!, inka

Subscribers: