[identity] identity service changes corresponding to proto updates
Summary: The identity service changes for the proto changes in D10657.
This will fail the CI because I haven't made any client changes in this diff, but I hit an [issue](https://linear.app/comm/issue/ENG-6472/protoc-gen-grpc-web-version-change-causing-big-differences-in-grpc-web) with the grpc-web codegen, so kind of forced to split this up.
I'll make sure the last diff in this stack passes the CI before landing any of these changes.

Depends on D10657

Test Plan: called the update and delete RPCs from native_rust_library and confirmed that I could change the password successfully and delete an account

Reviewers: bartek!

Subscribers: