[native][web][lib] unify keyserver transform
Summary:
The only difference was that `sessionID` does not exist on native, so excluding this undefined field or setting it as `null` is not an issue.
We need to store this in the same version in SQLite, and for backup and schema consistency unify the transform, which later in the stack is used for converting data to SQLite format.

It also reduces code duplication and makes things more consistent.

Note: after migrating everything to SQLite those transforms are not needed anymore, and then those will be needed only for transforming data to SQLite.

Test Plan: Test if persistence works the same (mostly on native, for web it's just moving code to `/lib`)

Reviewers: inka

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10754