[identity] Store social proof in users table
Summary:

Part 1 of [[ https://linear.app/comm/issue/ENG-6480/store-socialproof-in-users-table | ENG-6480 ]].
When wallet user is being registered, we should store the social proof in the users table.

Depends on D10762

Test Plan:

Called `LogInWalletUser` RPC and checked (`awslocal dynamodb scan`) that the social proof is stored in the users table.

Reviewers: varun!, michal, will

Subscribers: