[identity] Remove social proof from devices table
Summary:

Part 2 of [[ https://linear.app/comm/issue/ENG-6480/store-socialproof-in-users-table | ENG-6480 ]].
Since the social proof is stored in the users table, we don't need to store it in the devices table.

Depends on D10764

Test Plan:

Repeated test plan from D10764. Verified that the social proof is still retrieved correctly, especially in GetKeyserverKeys RPC (for which I changed the code).

Reviewers: varun!, will, michal

Subscribers: