[lib][web][native] Delete keyserver accounts when deleting identity account
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6502/delete-keyserver-accounts-when-deleting-identity-account | ENG-6502 ]]
We want to have a singe action for deleting identity account and keyserver accounts, and an action for deleting just a keyserver account (for when the user disconnects from a keyserver)

Test Plan:
Tested that it is possible to delete account from both web and native and no errors apprear. Tested that if callback returned from `useDeleteAccount` throws on web, an error message is displayed in the modal.
Tested that when the deletion appears to have been succeful, the user is removed from keyserver db

Reviewers:
varun, michal, ashoat
Subscribers: