[lib] Expose a function that allows reading deviceID, userID and accessToken
Summary:
We need the deviceID and userID when calling auth endpoint in connection handler. The identity context has access to the most recent values - because it consumes the emitter on native.

Depends on D10777

Test Plan: Console log the result of this function and check if it is correct on native and web.

Reviewers: varun, inka, kamil

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10786