[identity] Basic implementation of UpdateDeviceList RPC
Summary:
This implements small part of [[ https://linear.app/comm/issue/ENG-5101/new-rpc-to-update-devicelist | ENG-5101 ]],
which is necessary for [[ https://linear.app/comm/issue/ENG-5425/primary-device-sends-the-identity-service-the-updated-device-list | ENG-5425 ]] to work.

Only implemented adding new devices to the device list, but not removing them. Appropriate checks have been added.

Depends on D10798

Test Plan:
Called the RPC from the client and verified that the device list is updated in the database.
Also verified that the device list is not updated if the update is anything else than adding a new SINGLE device.

Reviewers: varun, will, michal

Reviewed By: varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10799