[native] consolidate staff context provider into one file
Summary: As I was building the build info page for web, I noticed that the `hasStaffUserLoggedIn` only lives in `native`. To get this field into `web` we should lift the staff context provider into `lib` so that it can be used on both platforms.

Test Plan: `flow` + confirmed that there were no regressions with `native`

Reviewers:

Subscribers: