Implement attachments capture
Summary: This differential adds method that parses logs in search for blob hashes and stores all found hashes to an attachments file.

Test Plan: The same as for parent differential but this time ensure that attachments file are created especially if you send multimedia messages. Attachments file is
a list of blob hashes separated by newline.

Reviewers: michal, kamil

Subscribers: