[web] convert pendingUsersToAdd into a map
Summary:
**PLEASE NOTE THAT THIS DIFF AND SUBSEQUENT DIFFS IN THIS STACK WILL NOT BE LANDED UNTIL MORE OF THE REDESIGN IS READY SINCE THIS WILL CAUSE REGRESSIONS IN PROD**

This next part of the stack will introduce the concept of "vip users" (happy to change the name of this just couldn't think of anything better). "vip users" will be selected users that are visually separated from the other users (similar to Linear). For context here is what the selected vip users section should look like:

{F1088612}

As I was implementing this I realized that `pendingUsersToAdd` should actually be a map and not a set. This diff makes that following change

Linear task: https://linear.app/comm/issue/ENG-5960/selected-users-section

Depends on D10747

Test Plan: `flow` + confirmed that there were no regressions to the add users experience

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10819