[lib] Logout after failing to auth with Ashoat's keyserver
Summary:
https://linear.app/comm/issue/ENG-6400/log-the-user-out-of-identity-if-login-to-ashoats-keyserver-fails

Depends on D10831

Test Plan: Throw an exception while logging in to the keyserver and check if a user gets logged out. At this point the action doesn't log out from identity (should be implemented as a part of https://linear.app/comm/issue/ENG-1075/have-clients-registerlogin-directly-with-identity-service-to-then-log) but in https://linear.app/comm/issue/ENG-6424/delete-state-on-identity-login-logout-delete-account we've decided it will be the same action which logs out from a keyserver.

Reviewers: varun, inka, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10834