[backup] Namespace log id in db
Summary: This was an idea that I wanted to implement but forgot before. We can prefix the `backupID` partitioning key in the log table with the `userID`. I think this is better because it allows us to map the log directly to the backup in the backup table (before we didn't knew the `userID` which is the partitioning key of the backup table). This is a quick change and we still aren't using backup service in production so I decided to implement it now. An alternative would be to add a new attribute but I prefer to keep them both in the partitioning key as they really represent a one things -> a specific backup of a given user. And if we ever need to iterate over specific user's logs we will also need to iterate over the compaction data so it wouldn't gain us anything.

Test Plan: Run backup integration tests.

Reviewers: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10857