[web] fix button in add friends modal not visible on small screens
Summary: The cta button in the add friend/block friend modal were not visible on smaller screens. This is because we were setting a fixed height that would extend above the screens. This diff fixes that bug + some copy that was requested in the linear task.

Test Plan: Please see the screenshots below

Reviewers:

Subscribers: