[native][web] Clear user specific data on identity actions
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6424/delete-state-on-identity-login-logout-delete-account | ENG-6424 ]]
Login will be handled once it's implemented. It's being tracked in [[ https://linear.app/comm/issue/ENG-6598/handle-login-action | ENG-6598 ]]
I will remove logic from specific reducers in next diffs
I had to change the dafault state on web - I need the `navInfo` to be correct. This is how we do it on native. I don't think it's the best, but at least the data is of correct shape, and again - this is how we do it on native, so I think we can use this for now

Test Plan: Tested that it is possible to log out and back in, and no errors show up. Tested that the data is cleared on logout success

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10860