[web] introduce member list sidebar provider
Summary:
The chat member list sidebar needs to be able to open + close from different points in the app. We will use `showMemberListSidebar` to track when the member list is open or close. Since we need to keep track of this state and use it many places, rather than prop drill this state to different places we should introduce a context provider that can be consumed in all the necessary parts

Linear task: https://linear.app/comm/issue/ENG-5975/introduce-new-member-list-section

Test Plan:
Confirmed that the chat member list sidebar can open + close (will introduce the logic for that in subsequent diffs)

{F1103269}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10899