[web] Rename `NavigationTab` -> `WebNavigationTab`
Summary:
Part of https://linear.app/comm/issue/ENG-6516/split-initialreduxstateresponse-into-client-and-server-variants

In order to move `InitialReduxStateResponse` to `lib`, we need to move `NavInfo` to `lib`.

In order to move `NavInfo` to `lib`, we need to move `NavigationTab`, `LoginMethod`, NavigationSettingsSection`, and `NavigationChatMode` to `lib.

Because those types are `web`-specific, we should give them a `Web*` prefix in `lib`. This is the first diff that does that for `NavigationTab`. Would've prefered to give them all the types the `Web*` prefix in one diff, but will break it up for easier review.

Test Plan: Straightforward rename, used IDE `Refactor -> Rename` tool.

Reviewers: ashoat, ginsu, tomek

Reviewed By: ginsu

Differential Revision: https://phab.comm.dev/D10886