[lib] Replace `ConnectionInfo.showDisconnectedBar` with `ConnectionInfo.unreachable`
Summary:
We want to rename `showDisconnectedBar` to be more general so it can be used for eg the keyserver status page.

We actually shouldn't need any migrations here since `KeyserverInfo.connection` is stripped out from `KeyserverInfo` before being persisted, and `KeyserverInfo.connection` is set to `defaultConnectionInfo` when being rehydrated. This is handled in `keyserverStoreTransform` on both `web` and `native`.

> Do we need to send "legacy" `ConnectionInfo`s to old clients?

My understanding is that `ConnectionInfo` is a client-only type and no inital value is sent from the `keyserver`. I was a bit confused because of `connectionInfoValidator`, but it turns out `connectionInfoValidator` is only consumed by `keyserverInfoValidator` which is only consumed by `keyserverStoreValidator`... which doesn't seem to be used anywhere? It looks like the only two KeyserverInfo fields we send as part of initialReduxState are `sessionID` and `updatesCurrentAsOf`.

---

Depends on D10879

Test Plan:
1. Open app
2. Ensure that `KeyserverStore` is as expected in Redux DevTools with `showDisconnectedBar` replaced by `unreachable`.
3. Flip on and off `keyserver` (killing and restarting) and ensure state updates as expected

{F1102762}
{F1102765}

---

Make sure old `native` clients continue to work as expected without any `validation` issues (not sure what specific flow to test since `connectionInfoValidator` appears unused?):

{F1102766}

Reviewers: ashoat, ginsu, tomek, inka

Reviewed By: tomek, inka

Differential Revision: https://phab.comm.dev/D10882