[lib] Add hook for easy access to CSAT
Summary:

Adds a convenience hook to get CSAT in React components. Ideally, it'd be best to get these from Redux / Context, but `getAuthMetadata()` is async so I need to use state + effect.
Let me know if there's a better way.

Depends on D10908.

Test Plan:

Flow. This is used extensively and tested in the next diff.

Reviewers: tomek!, inka!, varun, michal

Subscribers: