[identity] use siwe message and signature from primary device as social proof
Summary: we don't need the social_proof field in IdentityKeyInfo on wallet login. when we are registering a new wallet user, the SIWE message and signature are the social proof for the user. we can serialize this struct for now to avoid any DDB changes. in the future, we may want to store the signature and message separately, but @will can handle this as part of [ENG-6372](https://linear.app/comm/issue/ENG-6372/determine-proto-format-for-representing-a-users-username).

Test Plan: haven't tested yet, but in following diff will test by calling the login_wallet_user RPC from native and confirming that the serialized social proof is present in DDB

Reviewers: bartek, will, ashoat

Subscribers: tomek, will

Differential Revision: https://phab.comm.dev/D10902