[web] update toggleUser to use previouslySelectedUsers map if previously selected user is toggled
Summary:
**PLEASE NOTE THAT THIS DIFF AND SUBSEQUENT DIFFS IN THIS STACK WILL NOT BE LANDED UNTIL MORE OF THE REDESIGN IS READY SINCE THIS WILL CAUSE REGRESSIONS IN PROD**

As I was testing the new add users list, I ran into a case where the app would crash if I selected a user that I had searched for a user that was selected from the `userStoreSearchResults` cleared the search text, so this user was part of the vip section and then unselected that user and then reselected the user again

{F1088842}

This is because w/o the search text present `mergedUserInfos` does not have context regarding that searched user's user info. However, the `vipPendingUsers` still has the `GlobalAccountUserInfo` for that user stored.

We need to add a check to `toggleUser` that if the user toggled is a vip user, we should just grab the `GlobalAccountUserInfo` for that user from `vipPendingUsers` to prevent this crash from happening

Linear task: https://linear.app/comm/issue/ENG-5960/selected-users-section

Depends on D10822

Test Plan:
Please see demo video below

{F1088852}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10823