[web] introduce useUserRelationshipUserInfos
Summary:
**PLEASE NOTE THAT THIS DIFF AND SUBSEQUENT DIFFS IN THIS STACK WILL NOT BE LANDED UNTIL MORE OF THE REDESIGN IS READY SINCE THIS WILL CAUSE REGRESSIONS IN PROD**

In order to use the add users list for both add friends + add members, we need to make this component completely generic. Right now there is a lot of business logic for searching for any user to add as a friend that needs to be factored out.

As a first step to this I introduced `useUserRelationshipUserInfos` and moved all the search logic from `AddUsersList` into this new hook

Depends on D10849

Test Plan: `flow` + confirmed the add friends modal continues to work as expected

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10936