[web] introduce add users list provider
Summary: In order to use the add users list for both add friends + add members, we need to make this component completely generic. Right now there is a lot of business logic for searching for any user to add as a friend that needs to be factored out.

Test Plan: `flow` + confirmed the add friends modal continues to work as expected

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10937