[web] introduce add users list provider
Summary:
**PLEASE NOTE THAT THIS DIFF AND SUBSEQUENT DIFFS IN THIS STACK WILL NOT BE LANDED UNTIL MORE OF THE REDESIGN IS READY SINCE THIS WILL CAUSE REGRESSIONS IN PROD**

In order to use the add users list for both add friends + add members, we need to make this component completely generic. Right now there is a lot of business logic for searching for any user to add as a friend that needs to be factored out.

As I was attempting to lift the rest of the add friends specific logic from `AddUsersList`, I noticed that I needed to drill a ton of props in a bunch of places to handle the state needed for `AddUsersList`. This seemed a bit like a red flag to me, and I thought it would be a lot cleaner/better if I introduced a new provider to handle the state needed in `AddUsersList` and any relevant components

Depends on D10936

Test Plan: `flow` + confirmed the add friends modal continues to work as expected

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10937