[web] make AddUsersList feature agnostic
Summary:
**PLEASE NOTE THAT THIS DIFF AND SUBSEQUENT DIFFS IN THIS STACK WILL NOT BE LANDED UNTIL MORE OF THE REDESIGN IS READY SINCE THIS WILL CAUSE REGRESSIONS IN PROD**

In order to use the add users list for both add friends + add members, we need to make this component completely generic. Right now there is a lot of business logic for searching for any user to add as a friend that needs to be factored out.

This diff introduces a new generic `BaseAddUserInfo` type that both adding friends + adding members modals can use and lifts the rest of the business logic for adding friends out of `AddUsersList` and into `AddUsersListModalContent`

Depends on D10937

Test Plan: `flow` + confirmed the add friends modal continues to work as expected

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10940