[lib] Bump some endpoint timeouts in `thread-actions`
Summary:
Based on investigation in https://linear.app/comm/issue/ENG-6136/bump-http-timeouts-for-endpoints-as-needed, the following endpoints needed timeouts bumped:
- `delete_thread`
- `remove_members`
- `update_thread`
- `update_role`

Test Plan:
Made sure all of the endpoints continued to work as expected.

Trusting the the `*EndpointOptions` options function as expected and that timeouts will increase. Don't think it's necessary to try to get each of the endpoints to time out and log how long it took the time out. Could add this to Test Plan if deemed necessary.

Reviewers: ashoat, ginsu, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D10914