[services] update `Cargo.lock` in commtest
Summary:
Probably this change wasn't included in D10802.

To find things like this in the current cargo setup we need to run `cargo build` in all services.

Here is a task to fix the setup using Cargo workspace: [ENG-4630](https://linear.app/comm/issue/ENG-4630/consider-using-cargo-workspace-for-services).

Creating this diff because IDE finds issues with this and outdated Cargo locks shows up in my git changes.

Similar to D10366 and D10574.

Test Plan: N/A

Reviewers: will, varun, bartek

Reviewed By: will

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10915