[native] Bump `@redux-devtools/cli` to `^2.0.0`
Summary:
Bumping `@redux-devtools/cli` to `^2.0.0` resolves my issues with using Redux DevTools.

Changelog:
{F1106486}

Test Plan:
There are no patches associated with this dependency that need to be re-applied or tested.

After applying this change, I was able to use Redux DevTools for minutes at a time without losing any of the action entries:

{F1106564}
https://blob.sh/4dd3b4.png

Reviewers: ashoat, ginsu, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D10916