[lib] introduce useMembersGroupedByRole
Summary: Our new chat member list sidebar organizes the list of members in a chat by their role. This diff introduces a new hook that give a `threadInfo` will group all the members of that thread by their role and return an array of grouped role infos.

Test Plan: Confirmed that the roles + members for threads matched exaclty the same values that we had in our old thread members modal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10919