[lib] Extract action types constants out of socket-types.js
Summary:
In a later diff, we end up with an import cycle after adding an action type constant to this file and importing it from `CallKeyserverEndpointProvider`.

To avoid that import cycle, I decided to move the action types in this file to another file. I then modified the following diff to define the new action type constant in the new file.

Depends on D10946

Test Plan:
I tested this stack by testing the `CallKeyserverEndpoint` recovery mechanism in isolation:

1. I prevented the `Socket` from rendering by adding a `return null` line before the other returns in `KeyserverConnectionHandler`. This avoided the `Socket` triggering session recovery.
2. I started the iOS simulator and logged in using a test user.
3. I opened the Redux Dev Tools
4. I deleted the test user's cookie from the MariaDB database: `DELETE FROM cookies WHERE user = 6390578 AND platform = 'ios'`
5. I sent a message as the test user
6. I confirmed that session recovery was triggered in the Redux dev tools (and through some console logs)
7. I repeated the process above several times to make sure it consistently worked multiple times in a single run
8. I confirmed that the message was delivered "transparently" (without any visible issues, or evidence of session invalidation)

Reviewers: tomek, inka

Subscribers: