[lib] Add identityInvalidSessionDowngrade
Summary:
Adding a function that checkes if session downgrade is valid or not. Similarily to `invalidSessionDowngrade`, we check if
1. Some user is currently logged in
2. The action is trying to downgrade the session (currentUserInfo is being set to null or {anonymous: true})
3. The user is the same user the action was dispatched for
4. The session is the same session the action was dispatched for (checking if CSAT changed)

See [[https://linear.app/comm/issue/ENG-5888/refactor-invalidsessiondowngrade-[converted-to-refactor#comment-df3174a1 | this discussion]] for examples of scenarios in which a session downgrade is invalid

Test Plan:
Tested that if different user infos are provided in `actionCurrentUserInfo` and `preRequestUserState.currentUserInfo` then the function returns true.
Tested that if a different CSAT is present in `currentReduxState` and `preRequestUserState`, the function returns true.

Reviewers:
michal, kamil, ginsu
Subscribers: